add cross-compilation pipeline for 6 targets
Toolchain files in shared/toolchains/ for both Make (.mk) and CMake (.cmake):
  - aarch64-darwin: macOS Apple Silicon (native)
  - x86_64-darwin:  macOS Intel (cross via -arch x86_64)
  - x86_64-linux:   Linux AMD64/Intel 64 (x86_64-linux-gnu-gcc)
  - i686-linux:     Linux x86 32-bit (i686-linux-gnu-gcc)
  - aarch64-linux:  Linux ARM64 (aarch64-linux-gnu-gcc)
  - libnx:          Nintendo Switch (devkitPro devkitA64)

Build script: shared/build.sh <project> <target>
Setup script: shared/setup-toolchains.sh [target...]

Makefile updated to accept TOOLCHAIN= for cross-compilation
while preserving native auto-detection. Fixed sign-compare
warnings exposed by GCC (stricter than Apple Clang).

Verified: aarch64-darwin (Mach-O arm64), x86_64-darwin (Mach-O x86_64),
x86_64-linux (ELF x86-64) all build clean. Native tests: 35/35.

+++ src/pipe.c
@@ -373,7 +373,7 @@ pompom_pipe_t *pompom_pipe_accept(pompom_host_t *host,
     uint8_t f = buf[5];
     uint64_t total = pp_get64(buf + 6);
     uint16_t nlen = pp_get16(buf + 14);
-    if (16 + nlen > (size_t)n) return NULL;
+    if ((size_t)(16 + nlen) > (size_t)n) return NULL;
 
     pompom_pipe_t *p = calloc(1, sizeof(*p));
     if (!p) return NULL;
